requirements.txt
amqp==5.1.1
ansicon==1.89.0
anyio==3.6.1
arrow==1.2.2
asgiref==3.5.2
async-generator==1.10
asyncio==3.4.3
attrs==21.4.0
billiard==3.6.4.0
blessed==1.19.1
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.4
DateTime==4.5
Django==4.0.6
django-celery-results==2.4.0
django-picklefield==3.1
django-q==1.3.9
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
jinxed==1.2.0
kombu==5.2.4
outcome==1.2.0
prompt-toolkit==3.0.30
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==3.5.3
rfc3986==1.5.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wsproto==1.1.0
zope.interface==5.4.0